route for dashboard shows authenticated user header
